fix function formatting and moved components around and other random fixes
