+++ src/components/EmailSending.module.css
@@ -1,0 +1,5 @@
+.email-input {
+  border-color: black;
+  border-width: 1px;
+  margin: 5px;
+}
