--- src/components/READ.me
@@ -1,1 +1,0 @@
-Need so can push an empty directoy
