--- src/components/SendEmail.css
@@ -1,17 +1,0 @@
-.email-input {
-  border-color: black;
-  border-width: 1px;
-  margin: 5px 5px 5px 5px;
-}
-
-.submit-button {
-  background-color: #319795; /* Green */
-  border: none;
-  border-radius: 15%;
-  color: white;
-  padding: 10px 15px;
-  text-align: center;
-  display: inline-block;
-  font-size: 16px;
-  margin: 4px 2px;
-}
