+++ src/util/utils.js
@@ -1,31 +1,28 @@
 import axios from 'axios';
+import { renderEmail } from 'react-html-email';
 
 let baseURL = '';
 if (!process.env.NODE_ENV || process.env.NODE_ENV === 'development') {
   // dev code
-  baseURL = `${process.env.REACT_APP_API_URL}`;
+  baseURL = `${process.env.REACT_APP_HOST}:${process.env.REACT_APP_PORT}`;
 } else {
   // production code
   baseURL = `${process.env.REACT_APP_PROD_API_URL}`;
 }
 
 // See auth_utils for AuthInterceptor
-const PNPBackend = axios.create({
+export const PNPBackend = axios.create({
   baseURL,
   headers: { 'Access-Control-Allow-Credentials': '*' },
 });
 
-// export const sendEmail = async (name, email, messageHtml, subject) => {
-//   const response = await FYABackend.post('/nodemailer/send', {
-//     name,
-//     email,
-//     messageHtml: renderEmail(messageHtml),
-//     subject,
-//   });
-//   if (response.status !== 200) {
-//     throw new Error('Oops, something went wrong. Try again');
-//   }
-// };
-
-// eslint-disable-next-line import/prefer-default-export
-export { PNPBackend };
+export const sendEmail = async (newEmail, emailtemplate) => {
+  const response = await PNPBackend.post('/nodemailer/send', {
+    email: newEmail,
+    messageHtml: renderEmail(emailtemplate),
+    subject: 'test!',
+  });
+  if (response.status !== 200) {
+    throw new Error('Oops, something went wrong. Try again');
+  }
+};
